Introduce Bare mode to filter-schema transform
