+++ packages/transforms/filter-schema/yaml-config.graphql
@@ -1,3 +1,22 @@
 extend type Transform {
-  filterSchema: [String]
+  """
+  Transformer to filter (white/black list) GraphQL types, fields and arguments
+  """
+  filterSchema: FilterSchemaTransform
+}
+
+type FilterSchemaTransform @md {
+  """
+  Specify to apply filter-schema transforms to bare schema or by wrapping original schema
+  """
+  mode: FilterSchemaTransformMode
+  """
+  Array of filter rules
+  """
+  filters: [String!]!
+}
+
+enum FilterSchemaTransformMode {
+  bare
+  wrap
 }
