+++ packages/types/src/config-schema.json
@@ -337,11 +337,8 @@
           "$ref": "#/definitions/FederationTransform"
         },
         "filterSchema": {
-          "type": "array",
-          "items": {
-            "type": "string"
-          },
-          "additionalItems": false
+          "$ref": "#/definitions/FilterSchemaTransform",
+          "description": "Transformer to filter (white/black list) GraphQL types, fields and arguments"
         },
         "mock": {
           "$ref": "#/definitions/MockingConfig",
@@ -2021,6 +2018,27 @@
           "type": "string"
         }
       }
+    },
+    "FilterSchemaTransform": {
+      "additionalProperties": false,
+      "type": "object",
+      "title": "FilterSchemaTransform",
+      "properties": {
+        "mode": {
+          "type": "string",
+          "enum": ["bare", "wrap"],
+          "description": "Specify to apply filter-schema transforms to bare schema or by wrapping original schema (Allowed values: bare, wrap)"
+        },
+        "filters": {
+          "type": "array",
+          "items": {
+            "type": "string"
+          },
+          "additionalItems": false,
+          "description": "Array of filter rules"
+        }
+      },
+      "required": ["filters"]
     },
     "MockingConfig": {
       "additionalProperties": false,
